chore: fix quantization clippy warning

+++ crates/bitnet-quantization/src/i2s.rs
@@ -254,7 +254,7 @@ impl I2SQuantizer {
                 let i8_vec = _mm256_packs_epi16(i16_vec, i16_vec);
 
                 // Store 8 bytes
-                let result = _mm256_extract_epi64::<0>(i8_vec) as i64;
+                let result = _mm256_extract_epi64::<0>(i8_vec);
                 std::ptr::copy_nonoverlapping(
                     &result as *const i64 as *const i8,
                     output.as_mut_ptr().add(i * 8),
